requirements.txt
# KissBot V2 - Requirements pytwitchAPI
# Migration complète vers pytwitchAPI pour IRC natif ✅

# Bot Twitch - pytwitchAPI v4 (IRC + EventSub + Helix intégrés)
twitchAPI>=4.5.0

# HTTP et APIs
aiohttp==3.13.1
httpx==0.28.1
requests>=2.31.0

# Configuration
PyYAML==6.0.3

# Smart Context - Fuzzy matching pour noms de jeux
rapidfuzz>=3.0.0

# Monitoring système
psutil>=5.9.0

# Tests unitaires
pytest>=7.0.0
pytest-asyncio>=0.21.0

# INSTALLATION:
# pip install -r requirements.txt

# MIGRATION NOTES:
# - TwitchIO 2.x/3.x → pytwitchAPI 4.x
# - IRC natif sans bridges custom
# - Fonctionne partout (pas besoin de channel:bot scope)
# - Auto-reconnect et rate limiting intégrés